ATA-5199: connector: reorganization of metrics mid
